refactor: extract pure logic out of LocationManager/CameraScanner monoliths
No browser automation is available in this environment, so full UI
decomposition of these stateful components (visualizers, drawers) into
subcomponents carried too much unverifiable regression risk. Scoped this
pass to pure, verifiable extractions instead:

- utils/cardSort.js: the card sort-order comparator was copy-pasted 4x
  across autoSortContainerCards, findNextRecommendedSlot, the sorting
  assistant queue, and the unsorted list view. Consolidated into
  sortCardsByOrder(), verified against the original inline logic.
- utils/cardRarity.js: the rarity tier classification (border glow,
  badge color, badge label) was implemented twice in LocationManager.jsx.
  Consolidated and diffed byte-for-byte against the original inline
  boolean chains across a battery of rarity strings.
- utils/locationCoords.js: getPageNum/getSlotNum were redefined
  identically in 4 different functions.
- utils/pokemonTranslation.js + utils/formatPrice.js: pulled the JP->EN
  card-name dictionary/translator and price formatter out of
  CameraScanner.jsx (pure functions, no component state).

LocationManager.jsx: 3859 -> 3504 lines. CameraScanner.jsx: 1528 -> 1428
lines. Verified with production builds after each extraction plus
targeted equivalence tests (rarity classification and sort comparators
checked against the original logic; the JP-text regex's Unicode ranges
verified codepoint-for-codepoint against the \uXXXX escapes it replaced).
